+++ CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [6.x.x] - 2026-04-20
+## [7.x.x] - 2026-04-20
 
+- DEFEATURE: With `region` incoming, remove starting address from `output` statement.
 - FEATURE: added missing --list-extensions command line option which was already
   documented in the README.md
 - README: Removed obsolete text on if/else statement restrictions